Выносит генератор данных в отдельный модуль и подключает в точке входа
